+++ src/yrt_format_validator.mjs
